feat: 로그인 로직 개발

+++ src/main/java/com/cjw/reactivecommunityproject/common/security/model/SecurityAccessJwtVO.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Builder
 public record SecurityAccessJwtVO(
         String userUid,
-        Long roleUid
+        Integer roleUid
 ) implements Authentication, Serializable {
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
